Clean sliderValueParam to variable

+++ Tunny/Util/Variables.cs
@@ -2,12 +2,12 @@ namespace Tunny.Util
 {
     public struct Variable
     {
-        public decimal LowerBond { get; }
-        public decimal UpperBond { get; }
+        public double LowerBond { get; }
+        public double UpperBond { get; }
         public bool IsInteger { get; }
         public string NickName { get; }
 
-        public Variable(decimal lowerBond, decimal upperBond, bool isInteger, string nickName)
+        public Variable(double lowerBond, double upperBond, bool isInteger, string nickName)
         {
             LowerBond = lowerBond;
             UpperBond = upperBond;
